chore: added component VenueMoreInfos
